refactor: enhance signal initialization and subscription management
- Updated signal constructors across various components to accept a subscriptions map, improving flexibility and reducing redundancy in subscription handling.
- Refactored the trader initialization process to ensure the crypto trader is properly set up and reported as ready.
- Cleaned up obsolete code and improved the organization of signal management, enhancing maintainability and clarity in the codebase.
- Adjusted the measurement handling in the Thesis struct to streamline data retrieval and improve overall performance.

+++ types/measurement.go
@@ -168,7 +168,7 @@ type Measurement struct {
 	At           time.Time               `json:"at" validate:"required"`
 	ObservedFrom time.Time               `json:"observedFrom,omitempty"`
 	Horizon      time.Duration           `json:"horizon,omitempty" validate:"nonnegative"`
-	Maturity     int64                   `json:"maturity,omitempty" validate:"finite"`
+	Maturity     float64                 `json:"maturity,omitempty" validate:"finite"`
 	Uncertainty  *MeasurementUncertainty `json:"uncertainty,omitempty"`
 	Validity     MeasurementValidity     `json:"validity"`
 	Scale        ScaleReference          `json:"scale"`
